modularized rl_map initial
Initial commit of a modularized version of the functions for writing an rl_map from the CellPhoneDB reference and functions for conversion of genes in an rl_map to gene orthologs in another species. The ortholog conversion can handle one-many and many-one mappings. For one-many mappings, a seperate interaction is written for each possible ortholog. We are leaving this up to users that the vectors of "from" and "to" should be designed to have only high-confidence orthologs to avoid many new interactions being written in the rl_map with weak evidence.