new create domino initial
Generates a domino object, but has some outstanding issues. Quick tests with the inputs from sysdata. Cell barcode orders are not checked, so there is mismatch from the old output for the stored features matrix and results do not match for clust_de. There were also some NAs in the cor matrix I need to find the root cause of.