unique receptor names
Avoiding duplicate receptor names being pulled from rl.parse now ensures that the results for simple receptors matches the correlations assessed in 0.99.1. However, NAs are still present for a receptor present in the rl map, but not the expression matrices and  the complex receptor was not included in the cor matrix.